Move updateArray from observable-array to helpers
