+++ src/helpers.ts
@@ -329,3 +329,43 @@ export function entries(obj: any): [string, any][] {
 		resArray[i] = [ownProps[i], obj[ownProps[i]]];
 	return resArray;
 }
+
+function callSplice(arr: any[], index: number, removeCount: number, addItems: any[]): void {
+	if (removeCount || addItems.length > 0) {
+		var addItemsArgs = [index, removeCount].concat(addItems);
+		arr.splice.apply(arr, addItemsArgs);
+	}
+}
+
+export function updateArray(array: any[], values: any[]): void {
+	var pointer = 0;
+	var srcSeek = 0;
+	var tgtSeek = 0;
+
+	while (srcSeek < array.length) {
+		if (array[srcSeek] === values[tgtSeek]) {
+			if (pointer === srcSeek && pointer === tgtSeek) {
+				// items match, so advance
+				pointer = srcSeek = tgtSeek = pointer + 1;
+			}
+			else {
+				// remove range from source and add range from target
+				callSplice(array, pointer, srcSeek - pointer, values.slice(pointer, tgtSeek));
+
+				// reset to index follow target seek location since arrays match up to that point
+				pointer = srcSeek = tgtSeek = tgtSeek + 1;
+			}
+		}
+		else if (tgtSeek >= values.length) {
+			// reached the end of the target array, so advance the src pointer and test again
+			tgtSeek = pointer;
+			srcSeek += 1;
+		}
+		else {
+			// advance to the next target item to test
+			tgtSeek += 1;
+		}
+	}
+
+	callSplice(array, pointer, srcSeek - pointer, values.slice(pointer, Math.max(tgtSeek, values.length)));
+}
