+++ src/helpers.unit.ts
@@ -1,4 +1,4 @@
-import { merge } from "./helpers";
+import { merge, updateArray } from "./helpers";
 
 describe("merge", function () {
 	it("merges two input objects into a single new object", () => {
@@ -50,5 +50,20 @@ describe("merge", function () {
 
 		// Should return a new object
 		expect(result).not.toBe(obj1);
+	});
+});
+
+describe("updateArray", () => {
+	it.each([
+		[[], [1, 2, 3]], // Add items to the array
+		[[1, 2, 3], []], // Remove items from the array
+		[[1, 2, 3], [3, 2, 1]], // Reverse the array
+		[[1, 2, 3, 4], [3, 1, 4, 2]], // Change order of items in the array
+		[[1, 2, 3, 4, 5], [1, 4]], // Remove extra items
+		[[2, 5], [1, 2, 3, 4, 5]], // Add new items
+		[[1, 2, 3, 4, 5], [1, 3, 2, 4, 5]] // Move an item forward or back in the list
+	])("updates the target array to match the given input", (arr, values) => {
+		updateArray(arr, values);
+		expect(arr).toEqual(expect.arrayContaining(values));
 	});
 });
