+++ src/observable-array.unit.ts
@@ -1,5 +1,6 @@
 import { createEventObject } from "./events";
-import { ArrayChangeType, ObservableArray, updateArray } from "./observable-array";
+import { updateArray } from "./helpers";
+import { ArrayChangeType, ObservableArray } from "./observable-array";
 
 describe("ObservableArray", () => {
 	it("should not publish change event for no changes", () => {
@@ -197,20 +198,5 @@ describe("ObservableArray", () => {
 				]
 			}));
 		});
-	});
-});
-
-describe("updateArray", () => {
-	it.each([
-		[[], [1, 2, 3]], // Add items to the array
-		[[1, 2, 3], []], // Remove items from the array
-		[[1, 2, 3], [3, 2, 1]], // Reverse the array
-		[[1, 2, 3, 4], [3, 1, 4, 2]], // Change order of items in the array
-		[[1, 2, 3, 4, 5], [1, 4]], // Remove extra items
-		[[2, 5], [1, 2, 3, 4, 5]], // Add new items
-		[[1, 2, 3, 4, 5], [1, 3, 2, 4, 5]] // Move an item forward or back in the list
-	])("updates the target array to match the given input", (arr, values) => {
-		updateArray(arr, values);
-		expect(arr).toEqual(expect.arrayContaining(values));
 	});
 });
